# openapi.yaml
openapi: 3.1.0
info:
  title: pwrap control plane
  version: 0.1.0
  description: |
    Control plane API for pwrap. Manages projects, API keys, tenant migrations,
    branching, PostgREST token issuance, realtime subscriptions, and short-lived
    scoped DSN handoff for SDKs.

    Two auth surfaces:

    - **Bootstrap** (`PWRAP_BOOTSTRAP_TOKEN`) — root credential. Its only power
      is minting scoped admin tokens at `/v1/admin/tokens`; it is rejected on
      every other endpoint.
    - **Scoped admin token** (`pwa_…`) — management calls. Carries some of
      `projects`, `keys`, `migrate`, `sql`; each endpoint names the scope it
      needs. `sql` is separable because it executes arbitrary SQL as the tenant
      role.
    - **Project API key** (`pwk_…`) — the SDK surface. Exchanged at
      `/v1/connection` for a Postgres DSN, which the SDK then uses to talk to
      Postgres directly.

    Endpoints under `/v1/connection`, `/v1/rest/token` and `/v1/subscribe` are
    rate limited per API-key prefix (10 rps, burst 20).
servers:
  - url: http://localhost:8080
    description: local dev
paths:
...
  /v1/admin/tokens:
    post:
      summary: Mint a scoped admin token
      description: |
        Authenticated with the bootstrap token — a scoped token cannot mint, so
        it cannot widen its own privileges. The plaintext is returned once and is
        not recoverable afterwards.
      security: [{ BootstrapBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scopes]
              properties:
                name: { type: string, description: "human label, shown in the audit log" }
                scopes:
                  type: array
                  items: { type: string, enum: [projects, keys, migrate, sql] }
                  minItems: 1
                ttl_seconds: { type: integer, description: "0 or omitted = no expiry" }
      responses:
        "201": { description: issued, content: { application/json: { schema: { $ref: "#/components/schemas/IssuedAdminToken" } } } }
        "400": { description: "no scopes given, or an unknown scope" }
        "401": { description: missing or invalid bootstrap token }
    get:
      summary: List admin tokens (metadata only)
      security: [{ BootstrapBearer: [] }]
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AdminToken" }
  /v1/admin/tokens/{tokenID}:
    parameters:
      - { in: path, name: tokenID, required: true, schema: { type: string, format: uuid } }
    delete:
      summary: Revoke an admin token
      description: Takes effect on the token's next request.
      security: [{ BootstrapBearer: [] }]
      responses:
        "204": { description: revoked }
        "400": { description: bad token id }
        "404": { description: not found }
  /v1/projects:
    post:
      summary: Create project
      description: Provisions a Postgres role and schema (`p_<slug>`) for the project.
      security: [{ AdminBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties: { name: { type: string } }
      responses:
        "201": { description: created, content: { application/json: { schema: { $ref: "#/components/schemas/Project" } } } }
        "400": { description: invalid name }
        "409": { description: already exists }
    get:
      summary: List projects
      description: Requires the `projects` scope.
      security: [{ AdminBearer: [] }]
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Project" }
...
  /v1/connection:
    post:
      summary: Exchange API key for a Postgres DSN
      description: |
        SDK bootstrap endpoint. Authenticated with the project API key; the
        project is resolved from the key, so no id is in the path.

        **Credential lifetime.** `expires_at` is a *refresh hint* for the SDK,
        not an enforced expiry. The DSN carries the tenant role's stable
        password, so it keeps working after `expires_at` passes and after the
        issuing API key is revoked. Rotating per-issue credentials are planned;
        until then, treat an issued DSN as a long-lived secret and rotate the
        role password directly if one leaks.
      security: [{ APIKeyBearer: [] }]
      responses:
        "200":
          description: dsn minted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Connection" }
        "401": { description: missing or invalid key }
        "403": { description: revoked or expired }
        "429": { description: rate limited }
  /v1/rest/token:
    post:
      summary: Issue a PostgREST-compatible JWT
      description: |
        Returns an HS256 JWT whose `role` claim is the project's tenant role, for
        use against the PostgREST sidecar. Returns 503 when REST is not
        configured (`PWRAP_JWT_SECRET` / `PWRAP_AUTHENTICATOR_PASSWORD` unset).
      security: [{ APIKeyBearer: [] }]
      requestBody:
        description: optional
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:     { type: string, description: "echoed into request.jwt.claims for RLS" }
                ttl_seconds: { type: integer, description: "defaults to PWRAP_REST_TOKEN_TTL_SECONDS (3600)" }
      responses:
        "200": { description: issued, content: { application/json: { schema: { $ref: "#/components/schemas/RestToken" } } } }
        "401": { description: missing or invalid key }
        "404": { description: project not found }
        "429": { description: rate limited }
        "503": { description: REST integration disabled }
  /v1/subscribe:
    get:
      summary: Subscribe to change events (WebSocket)
      description: |
        Upgrades to a WebSocket and streams `ChangeEvent` JSON messages for the
        authenticated project's schema.

        Auth is via the `api_key` query parameter or an `Authorization` header.
        The query parameter exists because browsers cannot set headers on a
        `WebSocket` constructor — which also means the key appears in URLs and
        may be logged by proxies. Prefer the header wherever you control the
        client.
      parameters:
        - { in: query, name: api_key, schema: { type: string }, description: "required unless an Authorization header is sent" }
        - { in: query, name: table,   schema: { type: string }, description: "restrict to one table" }
        - { in: query, name: user_id, schema: { type: string }, description: "restrict to rows whose change carried this jwt user_id" }
      responses:
        "101": { description: switching protocols — the stream carries ChangeEvent messages }
        "401": { description: missing or invalid api key }
        "404": { description: project not found }
        "429": { description: rate limited }
components:
  securitySchemes:
    BootstrapBearer:
      type: http
      scheme: bearer
      description: >-
        PWRAP_BOOTSTRAP_TOKEN. Root credential, accepted ONLY on /v1/admin/tokens.
        Rejected everywhere else.
    AdminBearer:
      type: http
      scheme: bearer
      description: >-
        A scoped admin token (`pwa_…`) minted via POST /v1/admin/tokens. Each
        endpoint requires a specific scope; see its description.
    APIKeyBearer:
      type: http
      scheme: bearer
      description: "Project API key (pwk_...)"
  schemas:
...
    Project:
      type: object
      properties:
        id:                { type: string, format: uuid }
        name:              { type: string }
        slug:              { type: string }
        pg_role:           { type: string }
        pg_schema:         { type: string }
        parent_project_id: { type: string, format: uuid, nullable: true, description: "set when the project is a branch" }
        created_at:        { type: string, format: date-time }
        updated_at:        { type: string, format: date-time }
...
    AdminToken:
      type: object
      properties:
        id:           { type: string, format: uuid }
        prefix:       { type: string, description: "8-char lookup prefix; also what the audit log records" }
        name:         { type: string }
        scopes:
          type: array
          items: { type: string, enum: [projects, keys, migrate, sql] }
        last_used_at: { type: string, format: date-time, nullable: true }
        expires_at:   { type: string, format: date-time, nullable: true }
        revoked_at:   { type: string, format: date-time, nullable: true }
        created_at:   { type: string, format: date-time }
    IssuedAdminToken:
      allOf:
        - $ref: "#/components/schemas/AdminToken"
        - type: object
          required: [token]
          properties:
            token: { type: string, description: "plaintext (pwa_…), returned once" }
...
    Connection:
      type: object
      required: [dsn, schema, schema_version, expires_at]
      properties:
        dsn:            { type: string }
        schema:         { type: string }
        schema_version:
          type: string
          description: >-
            Latest migration version applied to the tenant schema; empty when
            migrations have never been applied. SDKs compare this against their
            compile-time constant and refuse a tenant that is behind.
        expires_at:
          type: string
          format: date-time
          description: "refresh hint, not an enforced expiry — see the endpoint description"
    RestToken:
      type: object
      required: [token, url, role, expires_at]
      properties:
        token:      { type: string, description: "HS256 JWT" }
        url:        { type: string, description: "PostgREST base URL (PWRAP_REST_URL)" }
        role:       { type: string, description: "Postgres tenant role encoded in the token" }
        expires_at: { type: string, format: date-time }
...